requirements.txt
anyio==3.7.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.5.7
cffi==1.15.1
cryptography==40.0.2
cycler==0.11.0
decorator==5.1.1
exceptiongroup==1.1.1
fonttools==4.33.3
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
kiwisolver==1.4.3
matplotlib==3.5.2
numpy==1.22.4
packaging==21.3
pandas==1.4.3
Pillow==9.1.1
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
python-telegram-bot==13.1
pytz==2022.1
six==1.16.0
sniffio==1.3.0
tornado==6.3.2
tzlocal==5.0.1
